refactor agent test (#9174)
